requirements.txt
asgiref==3.5.0
autopep8==1.6.0
dj-database-url==0.5.0
dj-static==0.0.6
Django==4.0.1
django-admin-list-filter-dropdown==1.0.3
django-jazzmin==2.4.9
django-on-heroku==1.1.2
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
gunicorn==20.1.0
psycopg2==2.9.3
psycopg2-binary==2.9.3
pycodestyle==2.8.0
PyJWT==2.3.0
pytz==2021.3
sqlparse==0.4.2
static3==0.7.0
toml==0.10.2
tzdata==2021.5
whitenoise==5.3.0
